requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
authlib==1.6.1
certifi==2025.1.31
cffi==1.17.1 
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6 
cryptography==45.0.5
cyclopts==3.22.5
dnspython==2.7.0
docstring-parser==0.17.0
docutils==0.22
email-validator==2.2.0
et-xmlfile==2.0.0
exceptiongroup==1.2.2
fastmcp==2.11.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.10.1
mdurl==0.1.2
more-itertools==10.7.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
parse==1.20.2
pathable==0.4.4
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.7.1
pygments==2.19.1
pyperclip==1.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rich==13.9.4
rich-rst==1.3.1
rpds-py==0.26.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.45.3
typer==0.16.0
typing-extensions==4.12.2
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.34.0
werkzeug==3.1.1
azure-storage-blob==12.26.0